+++ apps/api/src/common/filter/http-exception.filter.ts
@@ -1,7 +1,7 @@
 import { ExceptionFilter, Catch, ArgumentsHost, HttpException } from '@nestjs/common';
 import { Response } from 'express';
 
-import { ErrorResponse } from '@/response.interface';
+import { ErrorResponse } from '@/common/Interceptor/response.interface';
 
 @Catch(HttpException)
 export class HttpExceptionFilter implements ExceptionFilter {
